refactored functions to converters.rs
