+++ src/converters.rs
@@ -26,6 +26,7 @@
 */
 
 use std::any::Any;
+use std::cmp::min;
 use crate::slicers::FnLineBreak;
 
 pub(crate) mod self_converter;
@@ -45,5 +46,71 @@ pub(crate) trait Converter<'a> {
 pub trait ColumnBuilder {
     fn parse_value(&mut self, name: &[u8]) ->usize;
 
+}
+
+fn column_lenght_num_rightaligned(cursor: usize, data: &[u8], runes: i16) -> (usize, usize) {
+    let mut eat=data.iter();
+    let mut counted_runes=0;
+    let mut start:usize =0;
+    let mut stop:usize =min (data.len(), runes as usize);
+
+    while counted_runes< runes as usize {
+        let byten=    eat.nth(0);
+        let bb:u8=match byten {
+            None => {
+//TODO  we ran out of data,this is an error, fix later.
+                return (start,stop);
+            }
+            Some(b) => {
+                *b
+            }
+        };
+
+            match bb {
+                48..=57 =>{ return (start,stop)},
+                _ => {}
+            };
+        start+=1;
+        counted_runes+=1;
+    }
+
+    (start,stop)
+}
+
+fn column_lenght(cursor: usize, data: &[u8], runes: i16) -> usize {
+    let mut eat=data.iter();
+    let mut counted_runes=0;
+    let mut len:usize =0;
+    let mut units=1;
+
+        while counted_runes< runes as usize {
+
+            let byten=    eat.nth(units-1);
+
+            let bb:u8=match byten {
+                None => {
+                    return len;
+                }
+                Some(b) => {
+                    *b
+                }
+            };
+
+        units = match bb {
+            bb if bb >> 7 == 0 => 1,
+            bb if bb >> 5 == 0b110 =>  2,
+            bb if bb >> 4 == 0b1110 =>  3,
+            bb if bb >> 3 == 0b11110 => 4,
+            bb => {
+// TODO BAD ERROR HANDL
+                panic!("Incorrect UTF-8 sequence");
+            }
+        };
+
+        len+=units;
+        counted_runes+=1;
+    }
+
+    len
 }
